+++ operator_console/src/index.js
@@ -4,3 +4,8 @@ export { renderInvoiceStatement, NEXT_OPERATOR_ACTION } from "./invoice_statemen
 export { renderLineTable } from "./line_table.js";
 export { renderStatusChip, resolveStatementStatus } from "./status_chip.js";
 export { renderTenantPin } from "./tenant_pin.js";
+export {
+  renderCollectionCase,
+  COLLECTION_CUSTOMER_COPY,
+  nextOperatorActionCopy,
+} from "./collection_case.js";